QuickSort Listo
-se cambiaron los parametros tipo size_t en quicksort por int, ya que no se podia determinar cuando un indice salia del rango del arreglo para el ordenamiento.
-Quicksort ya ordena el dato abstraco elemento dentro de un std::vector<>

+++ Estructuras/Algoritmos.hpp
@@ -38,7 +38,7 @@
  * @param last  Indice a la posición final del arreglo
  */
 /* -------------------------------*/
-void QuickSort(std::vector<elemento*> elementos ,size_t first, size_t last);
+void QuickSort(std::vector<elemento*> elementos ,int first, int last);
 
 /* -------------------------------*/
 /**
